split off serialize into own files
