+++ merlict_c89/all_headers.h
@@ -42,6 +42,7 @@
 #include "mliDynMap.h"
 #include "mliBoundaryLayer.h"
 #include "mliMedium.h"
+#include "mliMedium_serialize.h"
 #include "mliMedium_json.h"
 #include "mliSurface.h"
 #include "mliFrame.h"
